# services/account/src/main/java/ru/abramov/practicum/bank/service/account/dto/ChangeBalanceDto.java
package ru.abramov.practicum.bank.service.account.dto;

import lombok.Data;

import java.math.BigDecimal;

@Data
public class ChangeBalanceDto {

    private BigDecimal amount;

    private Long version;
}